refactor(cli): remove commands.py compatibility shim
- migrate ~20 import sites to command_registry canonical path
- delete lazy _SlashCommandProxy forwarding module
- drop obsolete shim re-export test from test_command_registry
