+++ tests/cli/interactive_shell/sessions/test_resume_scenarios.py
@@ -10,7 +10,7 @@ from unittest.mock import patch
 import pytest
 from rich.console import Console
 
-from app.cli.interactive_shell.commands import dispatch_slash
+from app.cli.interactive_shell.command_registry import dispatch_slash
 from app.cli.interactive_shell.runtime.session import ReplSession
 from app.cli.interactive_shell.sessions.store import SessionStore
