+++ tests/cli/interactive_shell/test_command_registry.py
@@ -21,13 +21,7 @@ from app.cli.interactive_shell.command_registry.settings_cmds import (
     _VERBOSE_FIRST_ARGS,
 )
 from app.cli.interactive_shell.command_registry.tools_cmds import _TOOLS_FIRST_ARGS
-from app.cli.interactive_shell.commands import SLASH_COMMANDS as COMMANDS_EXPORT
 from app.cli.interactive_shell.runtime.session import ReplSession
-
-
-def test_commands_shim_reexports_same_registry() -> None:
-    assert COMMANDS_EXPORT["/help"] is SLASH_COMMANDS["/help"]
-    assert list(COMMANDS_EXPORT) == list(SLASH_COMMANDS)
 
 
 def _capture() -> tuple[Console, io.StringIO]:
